Make the test run only for SQL
